Further improvements to the convolution / correlation speed by using GEMM
Change the algorithm used to calculate the convolution and the correlation to one based on using the "gemm" blas operation. This is around 3 times faster than the previous "fast" algorithm for float and complex input tensors. For integer tensors this new algorithm is as fast as the previous algorithm. The reason it is not faster is that for integers we do not use BLAS' gemm function.

By using this new algorithm we were also able to add support for a new `down` argument for the `convolute` and `correlate` procedures. This will be useful to implement an efficient `upfirdn` function in the future.

This commit also adds many new tests for the `convolute` and `correlate` procedures. These are needed because to use the gemm based algorithm must handle differently the case in which the first input tensor is shorter than the second input tensor. Another set of tests are added because handling the different convolution / correlation modes using gemm is a bit tricky.

Thanks to @mratsim for the suggestion, and specially to @Vindaar for reviewing these changes and fixing multiple issues.