Improve fix for issue 637 (#654)
* Improve fix for issue #637 (https://github.com/mratsim/Arraymancer/issues/637)

The previous fix was incomplete. We could still assert if the input was an empty _tensor_.

This change also improves the handling of empty, rank-0 tensors. Up until now we would assert with the wrong message (talking about higher than rank-1 tensors not being supported). Now we don't assert and the code does the right thing.

* Improve the docs of `append` (use a code block for the examples)